2.1.1 pull request

+++ ita_root/ita_by_menu_export_import/backyard_main.py
@@ -210,11 +210,10 @@ def menu_import_exec(objdbca, record, workspace_id, workspace_path, uploadfiles_
                     shutil.rmtree(menu_dir)
 
         # 環境移行にて削除したテーブル名を記憶する用
-        deleted_table_list = []
+        imported_table_list = []
 
-        # 環境移行モードの場合はDELETE→INSERTするので特定のメニューは事前に退避しておく
-        if dp_mode == '1':
-            _dp_preparation(objdbca, workspace_id, menu_name_rest_list, execution_no_path, deleted_table_list)
+        # load_table.loadTableを使用するため特定のメニューは事前に処理しておく
+        _dp_preparation(objdbca, workspace_id, menu_name_rest_list, execution_no_path, imported_table_list, dp_mode)
 
         # 作成したview名を記憶する用
         tmp_table_list = []
@@ -248,7 +247,7 @@ def menu_import_exec(objdbca, record, workspace_id, workspace_path, uploadfiles_
                 # _dp_preparation()で既に処理していた場合はスキップする
                 continue
 
-            # 環境移行モードの場合、既存のデータは全て削除してからデータをインポートする
+            # 環境移行モードの場合、uploadfiles配下のデータを削除する
             if dp_mode == '1':
                 tmp_msg = "check information_schema.tables START: {}".format(table_name)
                 g.applogger.debug(addline_msg('{}'.format(tmp_msg)))  # noqa: F405
@@ -263,66 +262,66 @@ def menu_import_exec(objdbca, record, workspace_id, workspace_path, uploadfiles_
                     if os.path.isdir(uploadfiles_dir + '/' + menu_id):
                         shutil.rmtree(uploadfiles_dir + '/' + menu_id)
 
-                # DBデータファイル読み込み
-                db_data_path = execution_no_path + '/' + table_name + '.sql'
-                jnl_db_data_path = execution_no_path + '/' + table_name + '_JNL.sql'
+            # DBデータファイル読み込み
+            db_data_path = execution_no_path + '/' + table_name + '.sql'
+            jnl_db_data_path = execution_no_path + '/' + table_name + '_JNL.sql'
 
-                if table_name not in deleted_table_list:
-                    tmp_msg = "DROP and CREATE TABLE START: {}".format(db_data_path)
+            if table_name not in imported_table_list:
+                tmp_msg = "DROP and CREATE TABLE START: {}".format(db_data_path)
+                g.applogger.debug(addline_msg('{}'.format(tmp_msg)))  # noqa: F405
+
+                # テーブルを作成
+                objdbca.sqlfile_execute(db_data_path)
+
+                tmp_msg = "DROP and CREATE TABLE END: {}".format(db_data_path)
+                g.applogger.debug(addline_msg('{}'.format(tmp_msg)))  # noqa: F405
+
+                imported_table_list.append(table_name)
+                if os.path.isfile(jnl_db_data_path):
+                    tmp_msg = "DROP and CREATE JNL START: {}".format(jnl_db_data_path)
                     g.applogger.debug(addline_msg('{}'.format(tmp_msg)))  # noqa: F405
 
-                    # テーブルを作成
-                    objdbca.sqlfile_execute(db_data_path)
+                    # 履歴テーブルを作成
+                    objdbca.sqlfile_execute(jnl_db_data_path)
+                    imported_table_list.append(table_name + '_JNL')
 
-                    tmp_msg = "DROP and CREATE TABLE END: {}".format(db_data_path)
+                    tmp_msg = "DROP and CREATE JNL END: {}".format(jnl_db_data_path)
                     g.applogger.debug(addline_msg('{}'.format(tmp_msg)))  # noqa: F405
 
-                    deleted_table_list.append(table_name)
-                    if os.path.isfile(jnl_db_data_path):
-                        tmp_msg = "DROP and CREATE JNL START: {}".format(jnl_db_data_path)
+            if view_name:
+                if table_name.startswith('T_CMDB'):
+                    # 一度作成したview名は記憶する
+                    if view_name not in tmp_table_list:
+                        tmp_table_list.append(view_name)
+
+                        tmp_msg = "check information_schema.views START: {}".format(view_name)
                         g.applogger.debug(addline_msg('{}'.format(tmp_msg)))  # noqa: F405
 
-                        # 履歴テーブルを作成
-                        objdbca.sqlfile_execute(jnl_db_data_path)
-                        deleted_table_list.append(table_name + '_JNL')
+                        chk_view_sql = " SELECT TABLE_NAME FROM information_schema.views WHERE `TABLE_NAME` = %s "
+                        chk_view_rtn = objdbca.sql_execute(chk_view_sql, [view_name])
 
-                        tmp_msg = "DROP and CREATE JNL END: {}".format(jnl_db_data_path)
+                        tmp_msg = "check information_schema.views END: {}".format(view_name)
                         g.applogger.debug(addline_msg('{}'.format(tmp_msg)))  # noqa: F405
-
-                if view_name:
-                    if table_name.startswith('T_CMDB'):
-                        # 一度作成したview名は記憶する
-                        if view_name not in tmp_table_list:
-                            tmp_table_list.append(view_name)
-
-                            tmp_msg = "check information_schema.views START: {}".format(view_name)
+                        if len(chk_view_rtn) != 0:
+                            tmp_msg = "DROP VIEW START: {}".format(view_name)
                             g.applogger.debug(addline_msg('{}'.format(tmp_msg)))  # noqa: F405
 
-                            chk_view_sql = " SELECT TABLE_NAME FROM information_schema.views WHERE `TABLE_NAME` = %s "
-                            chk_view_rtn = objdbca.sql_execute(chk_view_sql, [view_name])
+                            drop_view_sql = "DROP VIEW IF EXISTS `{}`".format(view_name)
+                            objdbca.sql_execute(drop_view_sql, [])
 
-                            tmp_msg = "check information_schema.views END: {}".format(view_name)
+                            tmp_msg = "DROP VIEW END: {}".format(view_name)
                             g.applogger.debug(addline_msg('{}'.format(tmp_msg)))  # noqa: F405
-                            if len(chk_view_rtn) != 0:
-                                tmp_msg = "DROP VIEW START: {}".format(view_name)
-                                g.applogger.debug(addline_msg('{}'.format(tmp_msg)))  # noqa: F405
 
-                                drop_view_sql = "DROP VIEW IF EXISTS `{}`".format(view_name)
-                                objdbca.sql_execute(drop_view_sql, [])
+                        # DBデータファイル読み込み
+                        view_data_path = execution_no_path + '/' + view_name
+                        if os.path.isfile(view_data_path):
+                            tmp_msg = "CREATE VIEW START: {}".format(view_data_path)
+                            g.applogger.debug(addline_msg('{}'.format(tmp_msg)))  # noqa: F405
 
-                                tmp_msg = "DROP VIEW END: {}".format(view_name)
-                                g.applogger.debug(addline_msg('{}'.format(tmp_msg)))  # noqa: F405
+                            objdbca.sqlfile_execute(view_data_path)
 
-                            # DBデータファイル読み込み
-                            view_data_path = execution_no_path + '/' + view_name
-                            if os.path.isfile(view_data_path):
-                                tmp_msg = "CREATE VIEW START: {}".format(view_data_path)
-                                g.applogger.debug(addline_msg('{}'.format(tmp_msg)))  # noqa: F405
-
-                                objdbca.sqlfile_execute(view_data_path)
-
-                                tmp_msg = "CREATE VIEW END: {}".format(view_data_path)
-                                g.applogger.debug(addline_msg('{}'.format(tmp_msg)))  # noqa: F405
+                            tmp_msg = "CREATE VIEW END: {}".format(view_data_path)
+                            g.applogger.debug(addline_msg('{}'.format(tmp_msg)))  # noqa: F405
 
             objmenu = _register_data(objdbca, workspace_id, execution_no_path, menu_name_rest, menu_id, table_name)
             if history_table_flag == '1':
@@ -356,27 +355,27 @@ def menu_import_exec(objdbca, record, workspace_id, workspace_path, uploadfiles_
             os.remove(workspace_path + '/tmp/driver/import_menu/skip_all_service')
 
 
-def _dp_preparation(objdbca, workspace_id, menu_name_rest_list, execution_no_path, deleted_table_list):
+def _dp_preparation(objdbca, workspace_id, menu_name_rest_list, execution_no_path, imported_table_list, dp_mode):
     tmp_msg = '_dp_preparation START: '
     g.applogger.debug(addline_msg('{}'.format(tmp_msg)))  # noqa: F405
 
     if 'menu_list' in menu_name_rest_list:
-        _basic_table_preparation(objdbca, workspace_id, menu_name_rest_list, 'menu_list', execution_no_path, deleted_table_list)
+        _basic_table_preparation(objdbca, workspace_id, menu_name_rest_list, 'menu_list', execution_no_path, imported_table_list, dp_mode)
 
     if 'menu_table_link_list' in menu_name_rest_list:
-        _basic_table_preparation(objdbca, workspace_id, menu_name_rest_list, 'menu_table_link_list', execution_no_path, deleted_table_list)
+        _basic_table_preparation(objdbca, workspace_id, menu_name_rest_list, 'menu_table_link_list', execution_no_path, imported_table_list, dp_mode)
 
     if 'menu_column_link_list' in menu_name_rest_list:
-        _basic_table_preparation(objdbca, workspace_id, menu_name_rest_list, 'menu_column_link_list', execution_no_path, deleted_table_list)
+        _basic_table_preparation(objdbca, workspace_id, menu_name_rest_list, 'menu_column_link_list', execution_no_path, imported_table_list, dp_mode)
 
     if 'role_menu_link_list' in menu_name_rest_list:
-        _basic_table_preparation(objdbca, workspace_id, menu_name_rest_list, 'role_menu_link_list', execution_no_path, deleted_table_list)
+        _basic_table_preparation(objdbca, workspace_id, menu_name_rest_list, 'role_menu_link_list', execution_no_path, imported_table_list, dp_mode)
 
     tmp_msg = '_dp_preparation END: '
     g.applogger.debug(addline_msg('{}'.format(tmp_msg)))  # noqa: F405
 
 
-def _basic_table_preparation(objdbca, workspace_id, menu_name_rest_list, menu_name_rest, execution_no_path, deleted_table_list):
+def _basic_table_preparation(objdbca, workspace_id, menu_name_rest_list, menu_name_rest, execution_no_path, imported_table_list, dp_mode):
     tmp_msg = '_basic_table_preparation START: '
     g.applogger.debug(addline_msg('{}'.format(tmp_msg)))  # noqa: F405
 
@@ -384,14 +383,16 @@ def _basic_table_preparation(objdbca, workspace_id, menu_name_rest_list, menu_na
     objmenu = _create_objmenu(objdbca, menu_name_rest_list, menu_name_rest)
 
     menu_id, table_name, history_table_flag = _menu_data_file_read(menu_name_rest, 'menu_id', execution_no_path)
-    delete_sql = "DELETE FROM {}".format(table_name)
-    objdbca.sql_execute(delete_sql, [])
-    deleted_table_list.append(table_name)
-    _register_basic_data(objdbca, workspace_id, execution_no_path, menu_name_rest, table_name, objmenu=objmenu)
+    if dp_mode == '1':
+        delete_sql = "DELETE FROM {}".format(table_name)
+        objdbca.sql_execute(delete_sql, [])
+    imported_table_list.append(table_name)
+    _register_basic_data(objdbca, workspace_id, execution_no_path, menu_name_rest, menu_id, table_name, dp_mode, objmenu=objmenu)
     if history_table_flag == '1':
-        delete_jnl_sql = "DELETE FROM {}".format(table_name + '_JNL')
-        objdbca.sql_execute(delete_jnl_sql, [])
-        deleted_table_list.append(table_name + '_JNL')
+        if dp_mode == '1':
+            delete_jnl_sql = "DELETE FROM {}".format(table_name + '_JNL')
+            objdbca.sql_execute(delete_jnl_sql, [])
+        imported_table_list.append(table_name + '_JNL')
         _register_history_data(objdbca, objmenu, workspace_id, execution_no_path, menu_name_rest, menu_id, table_name)
     menu_name_rest_list.remove(menu_name_rest)
 
@@ -551,7 +552,9 @@ def _register_data(objdbca, workspace_id, execution_no_path, menu_name_rest, men
     return objmenu
 
 
-def _register_basic_data(objdbca, workspace_id, execution_no_path, menu_name_rest, table_name, objmenu=None):
+def _register_basic_data(objdbca, workspace_id, execution_no_path, menu_name_rest, menu_id, table_name, dp_mode, objmenu=None):
+    t_comn_menu_column_link = 'T_COMN_MENU_COLUMN_LINK'
+
     # DATAファイル読み込み
     sql_data = Path(execution_no_path + '/' + menu_name_rest).read_text(encoding='utf-8')
     json_sql_data = json.loads(sql_data)
@@ -569,6 +572,18 @@ def _register_basic_data(objdbca, workspace_id, execution_no_path, menu_name_res
 
     for json_record in json_sql_data:
         param = json_record['parameter']
+
+        # 移行先に主キーの重複データが既に存在するか確認
+        param_type = "Register"
+        if dp_mode == '2':
+            ret_t_comn_menu_column_link = objdbca.table_select(t_comn_menu_column_link, 'WHERE MENU_ID = %s AND COL_NAME = %s', [menu_id, pk])  # noqa: E501
+            pk_name = ret_t_comn_menu_column_link[0].get('COLUMN_NAME_REST')
+            pk_value = param[pk_name]
+            chk_pk_sql = " SELECT * FROM `" + table_name + "` WHERE `" + pk + "` = '" + pk_value + "'"
+            chk_pk_record = objdbca.sql_execute(chk_pk_sql, [])
+            if len(chk_pk_record) != 0:
+                # 主キーが既に存在する場合
+                param_type = "Update"
 
         # 最終更新者をバックヤードユーザに設定
         param['last_updated_user'] = g.get('USER_ID')
@@ -588,7 +603,10 @@ def _register_basic_data(objdbca, workspace_id, execution_no_path, menu_name_res
         g.applogger.debug(debug_msg)
         objdbca.db_transaction_start()
 
-        result = objdbca.table_insert(table_name, colname_parameter, pk)
+        if param_type == 'Register':
+            result = objdbca.table_insert(table_name, colname_parameter, pk)
+        elif param_type == 'Update':
+            result = objdbca.table_update(table_name, colname_parameter, pk)
 
         # コミット/トランザクション終了
         debug_msg = g.appmsg.get_log_message("BKY-20005", [])
@@ -846,7 +864,8 @@ def menu_export_exec(objdbca, record, workspace_id, export_menu_dir, uploadfiles
                     create_view_str = rec[0]['Create View']
                     # Create文から余計な文言を切り取る
                     end_pos = create_view_str.find(' VIEW ')
-                    create_view_str = create_view_str[:7] + create_view_str[end_pos:]
+                    create_view_str = create_view_str[:6] + create_view_str[end_pos:]
+                    create_view_str = create_view_str.replace('CREATE VIEW', 'CREATE VIEW IF NOT EXISTS')
                     view_data_path = dir_path + '/' + view_name
                     with open(view_data_path, "w") as f:
                         f.write(create_view_str)
@@ -863,6 +882,9 @@ def menu_export_exec(objdbca, record, workspace_id, export_menu_dir, uploadfiles
             sqldump_path = dir_path + '/' + table_name + '.sql'
 
             cmd = ["mysqldump", "--single-transaction", "--opt", "-u", db_user, "-p" + db_password, "-h", db_host, "--skip-column-statistics", db_database, "--no-data", table_name]
+            # 時刻指定の場合、DROP TABLE文を追加しないようオプションを追加
+            if mode == '2':
+                cmd.append("--skip-add-drop-table")
 
             sp_sqldump = subprocess.run(cmd, capture_output=True, text=True)
 
@@ -871,8 +893,13 @@ def menu_export_exec(objdbca, record, workspace_id, export_menu_dir, uploadfiles
                 log_msg_args = [msg]
                 api_msg_args = [msg]
                 raise AppException("499-00201", [log_msg_args], [api_msg_args])
+
+            sqldump_result = sp_sqldump.stdout
             with open(sqldump_path, 'w', encoding='utf-8') as f:
-                f.write(sp_sqldump.stdout)
+                if mode == '2':
+                    # 時刻指定の場合、dump結果を一部編集する
+                    sqldump_result = sqldump_result.replace('CREATE TABLE', 'CREATE TABLE IF NOT EXISTS')
+                f.write(sqldump_result)
 
         # インポート時に利用するT_COMN_MENU_DATAを作成する
         t_comn_menu_data_path = dir_path + '/T_COMN_MENU_DATA'
